backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
mysql-connector-python==8.0.32
bcrypt==4.0.1
python-dotenv==1.0.0
flasgger==0.9.5 